+++ src/components/common/global-navigation-bar/types.ts
@@ -1,0 +1,12 @@
+interface ItemWrapperProps {
+  currentTab?: boolean;
+}
+
+interface NavItemProps {
+  children: React.ReactNode;
+  href: string;
+  icon: React.ReactNode;
+  currentTab?: boolean;
+}
+
+export type { ItemWrapperProps, NavItemProps };
